Refactor structure of commands
