+++ lib/packwerk/commands/offense_printer.rb
@@ -2,26 +2,28 @@
 # typed: strict
 
 module Packwerk
-  module OffensePrinter
-    extend T::Sig
+  module Commands
+    module OffensePrinter
+      extend T::Sig
 
-    sig do
-      params(
-        offenses: T::Array[T.nilable(::Packwerk::Offense)],
-        out: StringIO,
-        style: T.any(T.class_of(OutputStyles::Plain), T.class_of(OutputStyles::Coloured))
-      ).void
-    end
-    def show_offenses(offenses, out, style)
-      if offenses.empty?
-        out.puts("No offenses detected 🎉")
-      else
-        offenses.each do |offense|
-          out.puts(offense.to_s(style))
+      sig do
+        params(
+          offenses: T::Array[T.nilable(Offense)],
+          out: StringIO,
+          style: T.any(T.class_of(OutputStyles::Plain), T.class_of(OutputStyles::Coloured))
+        ).void
+      end
+      def show_offenses(offenses, out, style)
+        if offenses.empty?
+          out.puts("No offenses detected 🎉")
+        else
+          offenses.each do |offense|
+            out.puts(offense.to_s(style))
+          end
+
+          offenses_string = Inflector.default.pluralize("offense", offenses.length)
+          out.puts("#{offenses.length} #{offenses_string} detected")
         end
-
-        offenses_string = Inflector.default.pluralize("offense", offenses.length)
-        out.puts("#{offenses.length} #{offenses_string} detected")
       end
     end
   end
